feat(shader-dsl): contextual literal lift in arithmetic/comparison — drop f32()/u32() wrappers
A bare number operand now types itself from the op's context instead of a global f32 default, so the
author drops the `f32()`/`u32()`/`i32()` wrapper in every arithmetic + comparison op:

  x.add(f32(0.5))   → x.add(0.5)        // f32 LHS → 0.5 lifts to f32
  u.lt(u32(1))      → u.lt(1)           // u32 LHS → 1 lifts to u32 (emits `1u`)
  i.eq(i32(0))      → i.eq(0)

The arithmetic path already did this for u32/i32 LHS (`bin`); the COMPARISON path was the gap — it lifted
every literal to f32, so a u32 compare needed the explicit `u32()`. Factored the shared `liftArg(o)` (number
→ this node's scalar for u32/i32, else the f32 lift) and routed `bin` + `cmp` through it. 168 wrappers
dropped across the shaders.

Byte-identical (the lift emits the same `0.5` / `1u` literal the wrapper did): tsc clean; full suite
6656 pass / 0 fail (every WGSL snapshot unchanged); real GPU OFM-bright renders. Numeric LITERALS only —
the `f32(expr)` CASTs (e.g. `f32(viewport.z)`) are untouched.

This is wave 1 of the literal-lift cleanup; vec constructors / min-max-clamp-mix / select args follow.

+++ src/shaders/compute-match.ts
@@ -71,7 +71,7 @@ export function matchComputeKernel(spec: MatchKernelSpec): ModuleDecl {
     const color = Var(vec4fT)
 
     // Parameterized if-else ladder — one arm per sorted pattern.
-    const chain = If(v.eq(f32(0)), () => {
+    const chain = If(v.eq(0), () => {
       assign(color, colorVec(armByPattern.get(sortedPatterns[0]!)!.colorHex))
     })
     for (let i = 1; i < sortedPatterns.length; i++) {
